Splits the ZeMainActivity into its composable screens
